+++ src/harlequin/components/history_screen.py
@@ -2,11 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, ClassVar
 
-from rich.columns import Columns
-from rich.console import Group, RenderableType
 from rich.padding import Padding
 from rich.style import Style
-from rich.text import Text
 from textual import on
 from textual.app import ComposeResult
 from textual.containers import Horizontal
@@ -34,40 +31,7 @@ class HistoryOption(Option):
     @property
     def prompt(self) -> RenderResult:
         """The prompt for the option."""
-        ts = self.item.executed_at.strftime("%a, %b %d %H:%M:%S")
-        if self.item.result_row_count < 0:
-            result = Text("ERROR", style=self.error_style, justify="right")
-        else:
-            res = (
-                f"{self.item.result_row_count:n} "
-                f"{'record' if self.item.result_row_count == 1 else 'records'}"
-                if self.item.result_row_count
-                else "SUCCESS"
-            )
-            elapsed = f"{self.item.elapsed:.2f}s"
-            result = Text.assemble(
-                (res, "bold"), " in ", (elapsed, "bold"), justify="right"
-            )
-        query_lines = self.item.query_text.strip().splitlines()
-        if len(query_lines) > 8:
-            continuation: RenderableType = Text(
-                f"… ({len(query_lines) - 7} more lines)\n", style="italic"
-            )
-            query_lines = query_lines[0:7]
-        else:
-            continuation = ""
-
-        return Padding(
-            Group(
-                Columns(
-                    renderables=[Text(ts, style="bold"), result],
-                    expand=True,
-                ),
-                "\n".join(query_lines),
-                continuation,
-            ),
-            pad=self.PADDING,
-        )
+        return Padding(self.item.render(error_style=self.error_style), pad=self.PADDING)
 
     def __rich__(self) -> RenderResult:
         return self.prompt
@@ -104,9 +68,7 @@ class HistoryScreen(Screen[str]):
             color=error_style.color, italic=error_style.italic, bold=error_style.bold
         )
         self.list = HistoryList(
-            *reversed(
-                [HistoryOption(q, error_style=error_text_style) for q in self.history]
-            )
+            *[HistoryOption(q, error_style=error_text_style) for q in self.history]
         )
         self.preview = TextEditor(
             language="sql", theme=self.theme, read_only=True, use_system_clipboard=False